Define module to enable the start of the game
The playGame function attaches an event listener to each block.
Within each event listener's event handler, the data attribute:
block's value is pushed to the array referenced by the variable:
selections. The purpose of the selections to ensure that the
value of selections are widely available within the Game Module.
For each event handler, the function displayChoice is invoked
in which the player's selection is displayed on the chart
and the status of the player's turn is changed.