Define the function getChoice to display results
PlayerX will be first to play the game and then PlayerO
will be next. We take the array referenced by the
variable: allSelections, which represents all the selections
made by playerX and playerO. In the array referenced by the
variable: allSelections, the elements that are even-indexed
represent Player X's selections and the elements that are
odd-indexed represent Player O's selections.

Upon a player clicking on a grid, the function playGame
pushes the data-attribute: block's value to the array
referenced by the variable: allSelections. It's important
to note that the variable: allSelections is widely available
to all functions within the Game module.

Based on the most recently added element in the array
referenced by the variable allSelections,the
function getChoice extracts the most recently
added element's index and the element in allSelections.
If the element's index is even, then we push the element
to object: playerX's property: choice. If the element's
index is odd, then we push the element to object:
playerO's property: choice.

Then, the player's choice is added to the block and
updates the turn status.