Define helper functions to play game
When either player has three selections, the helper function:
checkVictory checks if there are any winning selections. If
the player's selection matches with any winning selections,
then it inputs the player's selection on the grid and updates
the status as to who won the game.

If not, it will be the next player's turn and continues
the game. In order for a draw to occur, playerX
(also the first person to start) must have its selections
checked by the helper function: checkVictory. If checkVictory
returns false and the playerX's selection is exactly 5,
then it will be considered a draw and the program will
make the updates accordingly.