Refractor function that attaches event listeners
The current function playGame contains a chunk of code
that attaches an event listener for each block within
the gameboard. It was recommended not to use
a for loop to attach event listeners to multiple blocks
due to performance issues. Thus, the forEach method
was used to replace the for loop.