Define function to prevent players from selecting
If either player win or if the game will become a draw,
then the function disableChoices will prevent players
from clicking blocks or making any changes.

In the function disableChoices, the old_gameboard is cloned
meaning that the tree and subtree of old_gameboard is deep
copied and stored to the new variable: new_gameboard. When
replacing the old_gameboard with the new_gameboard, the trees
and subtrees of new_gameboard no longer contains event
listeners.